Merge in SIBO/sina from SIBO-980-unified_query to develop

* commit 'b5078e193adc1fee1aee4acbc3aa0b699437d4e5':
  SIBO-980: Fix up some comments
  Fix flake8 issues in notebook
  SIBO-908: Give find() a more prominent place in docs
  SIBO-980: Add documentation for find()
  SIBO-980: Hide DAO find() for cleanliness
  SIBO-980: Improve find() documentation and tests
  SIBO-980: Fix pylint issues, query_order tests, and a cassandra issue
  SIBO-980: Add first pass on the find() query
  SIBO-980: fix variable names for clarity
  SIBO-980: Add id_pool to _do_data_query()
  SIBO-994: fix typo
  Fix a few cassandra bugs
  SIBO-994: expand uri selection in cass and fix some broken tests
  SIBO-994: Fix variable names in tests
  SIBO-994: add has_any and has_all support to uri query for sql
  SIBO-991: fix type query for generators
  SIBO-991: expand type query in cass
  SIBO-991: expand type query in sql

+++ python/sina/datastores/cass.py
@@ -4,6 +4,7 @@ import numbers
 import logging
 # Used for temporary implementation of LIKE-ish functionality
 import fnmatch
+import itertools
 from collections import defaultdict
 
 import six
@@ -269,6 +270,8 @@ class RecordDAO(dao.RecordDAO):
         for partition, data_list in six.iteritems(from_scalar_list_batch):
             with BatchQuery() as batch_query:
                 for entry in data_list:
+                    if not entry[0]:
+                        continue  # Empty lists should not be inserted, they can't be queried.
                     # We track only the min and max, that's all we need for queries
                     table.batch(batch_query).create(name=partition,
                                                     min=min(entry[0]),
@@ -398,10 +401,16 @@ class RecordDAO(dao.RecordDAO):
         all the files, etc. We add those deletions to a batchquery for later
         execution.
 
+        Note that if we fail to find the record in the Records table, we assume
+        it doesn't exist and abort our attempt to delete it.
+
         :param batch: the batch to add the deletions to
         :param record_id: the id of the record we're deleting
         """
         LOGGER.debug("Generating batch deletion commands for %s", record_id)
+        if not self._one_exists(record_id):
+            LOGGER.debug("Record with id %s not found. Aborting deletion.", record_id)
+            return
         record = self.get(record_id)
         # Delete from the record table itself
         schema.Record.objects(id=record_id).batch(batch).delete()
@@ -448,46 +457,38 @@ class RecordDAO(dao.RecordDAO):
         """
         Given a list of Records, update them in the backend.
 
+        This is not ACID-safe!
+
         :param records: A list of Records to update.
         """
+        # Try to at least minimize the chance of a bad deletion.
+        for record in records:
+            is_valid, warnings = record.is_valid()
+            if not is_valid:
+                raise ValueError(warnings)
         self.delete(record.id for record in records)
         self.insert(records)
 
-    def data_query(self, **kwargs):
+    def _do_data_query(self, criteria, id_pool=None):
         """
-        Return the ids of all Records whose data fulfill some criteria.
+        Handle the backend-specific logic for the dao data_query.
 
-        Criteria are expressed as keyword arguments. Each keyword
-        is the name of an entry in a Record's data field, and it's set
-        equal to either a single value or a DataRange (see utils.DataRanges
-        for more info) that expresses the desired value/range of values.
-        All criteria must be satisfied for an ID to be returned:
-
-            # Return ids of all Records with a volume of 12, a quadrant of
-            # "NW", AND a max_height >=30 and <40.
-            data_query(volume=12, quadrant="NW", max_height=DataRange(30,40))
-
-        :param kwargs: Pairs of the names of data and the criteria that data
-                         must fulfill.
+        :param criteria: Dict of {data_name: criteria_to_fulfill}
+        :param id_pool: List of ids to restrict results to.
         :returns: A generator of Record ids that fulfill all criteria.
 
         :raises ValueError: if not supplied at least one criterion or given
                             a criterion it does not support.
         """
-        LOGGER.debug('Finding all records fulfilling criteria: %s', kwargs.items())
-        # No kwargs is bad usage. Bad kwargs are caught in sort_and_standardize_criteria().
-        # We will always have at least one entry in one of scalar, string, scalarlist, etc.
-        if not kwargs.items():
-            raise ValueError("You must supply at least one criterion.")
         (scalar, string, scalarlist,
-         stringlist, universal) = utils.sort_and_standardize_criteria(kwargs)
+         stringlist, universal) = utils.sort_and_standardize_criteria(criteria)
         result_ids = []
 
         # String and scalar values can be passed directly to _apply_ranges_to_query
-        for criteria, table_type in ((scalar, "scalar"),
-                                     (string, "string")):
-            if criteria:
-                result_ids.append(self._apply_ranges_to_query(criteria,
+        for criteria_group, table_type in ((scalar, "scalar"),
+                                           (string, "string")):
+            if criteria_group:
+                result_ids.append(self._apply_ranges_to_query(criteria_group,
                                                               table_type))
 
         # List values have special logic per type
@@ -503,6 +504,9 @@ class RecordDAO(dao.RecordDAO):
         # Universal values come in only one type for now.
         if universal:
             result_ids.append(self._universal_query(universal))
+
+        if id_pool is not None:
+            result_ids.append(id_pool)
 
         # If we have more than one set of data, we need to find the intersect.
         for id in utils.intersect_lists(result_ids):
@@ -780,26 +784,40 @@ class RecordDAO(dao.RecordDAO):
             for result in results:
                 yield model.generate_record_from_json(json_input=json.loads(result.raw))
 
-    def get_all_of_type(self, type, ids_only=False):
+    def _do_get_all_of_type(self, types, ids_only=False, id_pool=None):
         """
-        Given a type of Record, return all Records of that type.
+        Given an iterable of types of Record, return all Records of those types.
 
-        :param type: The type of Record to return, ex: run
+        :param types: An iterable of types of Records to return
         :param ids_only: whether to return only the ids of matching Records
                          (used for further filtering)
 
         :returns: A generator of Records of that type or (if ids_only) a
                   generator of their ids
         """
-        LOGGER.debug('Getting all records of type %s.', type)
         # There's an index on type, which should be alright as type is expected
         # to have low cardinality. If speed becomes an issue, a dedicated query table might help.
-        query = (schema.Record.objects.filter(type=type).values_list('id', flat=True))
+
+        # Cassandra won't let us do an in_ on type, so we chain (a Record can only have
+        # one type, so we're guaranteed no duplicates).
+        generators = []
+        for type in types:
+            query = schema.Record.objects.filter(type=type).values_list('id', flat=True)
+            generators.append(query)
+
+        # Cassandra also doesn't want to let us filter in_ on record ID currently
+        # (though this may change in the future), and so:
+        if id_pool is not None:
+            id_pool = set(id_pool)
+            match_ids = (x for x in itertools.chain(*generators) if x in id_pool)
+        else:
+            match_ids = itertools.chain(*generators)
+
         if ids_only:
-            for id in query:
+            for id in match_ids:
                 yield str(id)
         else:
-            for record in self.get(query):
+            for record in self.get(match_ids):
                 yield record
 
     def get_with_curve_set(self, curve_set_name, ids_only=False):
@@ -855,7 +873,7 @@ class RecordDAO(dao.RecordDAO):
 
         query_tables = [type_name_to_tables[type] for type in data_types]
 
-        ids = list(self.get_all_of_type(record_type, ids_only=True))
+        ids = list(self._do_get_all_of_type([record_type], ids_only=True))
 
         for query_table in query_tables:
             results = set(query_table.objects
@@ -864,37 +882,17 @@ class RecordDAO(dao.RecordDAO):
             for result in results:
                 yield result
 
-    def get_given_document_uri(self, uri, accepted_ids_list=None, ids_only=False):
+    @staticmethod
+    def _get_records_for_some_uri(uri, id_pool=None):
         """
-        Return all records associated with documents whose uris match some arg.
+        Handle the logic for a single URI in _do_get_given_document_uri.
 
-        Temporary implementation. THIS IS A VERY SLOW, BRUTE-FORCE STRATEGY!
-        You use it at your own risk! Do not expect it to be performant or
-        particularly stable. To avoid duplications, while it does return
-        a generator, it actually returns a generator of a set (so there's
-        no memory conserved).
-
-        Supports the use of % as a wildcard character.
-
-        :param uri: The uri to use as a search term, such as "foo.png"
-        :param accepted_ids_list: A list of ids to restrict the search to.
-                                  If not provided, all ids will be used.
-        :param ids_only: whether to return only the ids of matching Records
-                         (used for further filtering)
-
-        :returns: A generator of unique found records or (if ids_only) a
-                  generator of their ids.
+        Returns a generator of all records that match the uri.
         """
-        LOGGER.debug('Getting all records related to uri=%s.', uri)
-        if accepted_ids_list:
-            LOGGER.debug('Restricting to %i ids.', len(accepted_ids_list))
-        LOGGER.warning('Temporary implementation of getting Records based on '
-                       'Document URI. This is a very slow, brute-force '
-                       'strategy.')
-        if accepted_ids_list is not None:
-            base_query = schema.DocumentFromRecord.objects.filter(id__in=accepted_ids_list)
+        if id_pool is not None:
+            base_query = schema.DocumentFromRecord.objects.filter(id__in=id_pool)
         else:
-            # If we haven't had an accepted_ids_list passed, any filtering we do
+            # If we haven't had an id_pool passed, any filtering we do
             # will NOT include the partition key, so we need to allow filtering.
             base_query = schema.DocumentFromRecord.objects.allow_filtering()
         # If there's a wildcard
@@ -925,6 +923,51 @@ class RecordDAO(dao.RecordDAO):
         else:
             match_ids = base_query.filter(uri=uri).values_list('id', flat=True)
 
+        return match_ids
+
+    def _do_get_given_document_uri(self, uri, id_pool=None, ids_only=False):
+        """
+        Return all records associated with documents whose uris match some arg.
+
+        Temporary implementation. THIS IS A VERY SLOW, BRUTE-FORCE STRATEGY!
+        You use it at your own risk! Do not expect it to be performant or
+        particularly stable. To avoid duplications, while it does return
+        a generator, it actually returns a generator of a set (so there's
+        no memory conserved).
+
+        Supports the use of % as a wildcard character.
+
+        :param uri: The uri to use as a search term, such as "foo.png"
+        :param id_pool: A list of ids to restrict the search to. If not provided,
+                        all ids will be used.
+        :param ids_only: whether to return only the ids of matching Records
+                         (used for further filtering)
+
+        :returns: A generator of unique found records or (if ids_only) a
+                  generator of their ids.
+        """
+        LOGGER.debug('Getting all records related to uri=%s.', uri)
+        if id_pool:
+            LOGGER.debug('Restricting to %i ids.', len(id_pool))
+        LOGGER.warning('Temporary implementation of getting Records based on '
+                       'Document URI. This is a very slow, brute-force '
+                       'strategy.')
+        if isinstance(uri, utils.StringListCriteria):
+            generators = []
+            for criterion_entry in uri.value:
+                generators.append(self._get_records_for_some_uri(criterion_entry,
+                                                                 id_pool))
+            if uri.operation == utils.ListQueryOperation.HAS_ALL:
+                # While a natural choice would be to ORDER_BY and use utils.intersect_ordered,
+                # Cassandra currently (June 2021) doesn't support ordering by partition key
+                # OR ordering by clustering key when partition key is unrestricted, meaning
+                # we can't do that here. Hopefully we don't run into memory issues.
+                match_ids = (x for x in set.intersection(*[set(x) for x in generators]))
+
+            else:
+                match_ids = itertools.chain(*generators)
+        else:
+            match_ids = self._get_records_for_some_uri(uri, id_pool)
         if ids_only:
             for id in set(match_ids):
                 yield id
